Refactor: Relocate common UI components and utilities
- Moved common UI components (TopBar, BottomNavBar, ListItem, etc.) from `features.common` to `core.ui.baseComponents` and `core.ui.components`.
- Relocated utility functions (CollectEffect, DateFormatter) to `core.ui.utils`.
- Updated import paths across various feature modules to reflect the new locations.
- Renamed `getOrError` to `getOrThrow` in `data.remote.utils`.
- Added KDocs to API models and some UI components for better clarity.
- Removed unused code from `Theme.kt`, `Type.kt` and `NetworkModule.kt`.
