+++ app/src/main/java/com/backcube/economyapp/core/ui/components/ListItem.kt
@@ -1,4 +1,4 @@
-package com.backcube.economyapp.features.common.ui
+package com.backcube.economyapp.core.ui.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
