+++ app/src/main/java/com/backcube/economyapp/core/ui/components/ShowAlertDialog.kt
@@ -1,4 +1,4 @@
-package com.backcube.economyapp.features.common.ui
+package com.backcube.economyapp.core.ui.components
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ButtonColors
