+++ app/src/main/java/com/backcube/economyapp/core/ui/components/ShowProgressIndicator.kt
@@ -1,4 +1,4 @@
-package com.backcube.economyapp.features.common.ui
+package com.backcube.economyapp.core.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
