+++ app/src/main/java/com/backcube/economyapp/core/ui/components/date/DateMode.kt
@@ -1,0 +1,5 @@
+package com.backcube.economyapp.core.ui.components.date
+
+enum class DateMode {
+    START, END
+}
