+++ app/src/main/java/com/backcube/economyapp/core/ui/utils/CollectEffect.kt
@@ -1,4 +1,4 @@
-package com.backcube.economyapp.features.common.utils
+package com.backcube.economyapp.core.ui.utils
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
