+++ app/src/main/java/com/backcube/economyapp/data/remote/di/NetworkModule.kt
@@ -69,5 +69,4 @@ class NetworkModule {
         return NetworkConnectivityObserver(context)
     }
 
-
 }
