+++ app/src/main/java/com/backcube/economyapp/data/remote/models/request/accounts/AccountCreateRequestApiModel.kt
@@ -2,6 +2,13 @@ package com.backcube.economyapp.data.remote.models.request.accounts
 
 import com.backcube.economyapp.domain.models.accounts.AccountCreateRequestModel
 
+/**
+ * Модель запроса на создание нового счета
+ *
+ * @property name Название счета
+ * @property balance Начальный баланс счета
+ * @property currency Валюта счета (например, "RUB", "USD")
+ */
 data class AccountCreateRequestApiModel(
     val name: String,
     val balance: String,
