+++ app/src/main/java/com/backcube/economyapp/data/remote/models/response/accounts/AccountStateApiModel.kt
@@ -3,6 +3,14 @@ package com.backcube.economyapp.data.remote.models.response.accounts
 import com.backcube.economyapp.domain.models.accounts.AccountStateModel
 import com.backcube.economyapp.domain.utils.toCurrencyIsoCode
 
+/**
+ * Модель состояния счета (используется в истории изменений)
+ *
+ * @property id ID состояния
+ * @property name Название счета
+ * @property balance Баланс счета
+ * @property currency Валюта счета
+ */
 data class AccountStateApiModel(
     val id: Int,
     val name: String,
