+++ app/src/main/java/com/backcube/economyapp/features/account/AccountScreenRoot.kt
@@ -18,16 +18,16 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import com.backcube.economyapp.R
+import com.backcube.economyapp.core.ui.baseComponents.CustomTopBar
+import com.backcube.economyapp.core.ui.components.CustomFloatingButton
+import com.backcube.economyapp.core.ui.components.CustomListItem
+import com.backcube.economyapp.core.ui.components.ShowAlertDialog
+import com.backcube.economyapp.core.ui.components.ShowProgressIndicator
+import com.backcube.economyapp.core.ui.utils.CollectEffect
 import com.backcube.economyapp.domain.utils.formatAsWholeThousands
 import com.backcube.economyapp.features.account.store.models.AccountEffect
 import com.backcube.economyapp.features.account.store.models.AccountIntent
 import com.backcube.economyapp.features.account.store.models.AccountState
-import com.backcube.economyapp.features.common.baseComponents.CustomTopBar
-import com.backcube.economyapp.features.common.ui.CustomFloatingButton
-import com.backcube.economyapp.features.common.ui.CustomListItem
-import com.backcube.economyapp.features.common.ui.ShowAlertDialog
-import com.backcube.economyapp.features.common.ui.ShowProgressIndicator
-import com.backcube.economyapp.features.common.utils.CollectEffect
 import com.backcube.economyapp.ui.theme.LightGreen
 import com.backcube.economyapp.ui.theme.White
 import kotlinx.coroutines.flow.Flow
