--- app/src/main/java/com/backcube/economyapp/features/common/ui/date/DateMode.kt
@@ -1,5 +1,0 @@
-package com.backcube.economyapp.features.common.ui.date
-
-enum class DateMode {
-    START, END
-}
