+++ app/src/main/java/com/backcube/economyapp/features/income/IncomeScreenRoot.kt
@@ -21,14 +21,14 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import com.backcube.economyapp.R
 import com.backcube.economyapp.core.navigation.Screens
+import com.backcube.economyapp.core.ui.baseComponents.CustomTopBar
+import com.backcube.economyapp.core.ui.components.CustomFloatingButton
+import com.backcube.economyapp.core.ui.components.CustomListItem
+import com.backcube.economyapp.core.ui.components.ShowAlertDialog
+import com.backcube.economyapp.core.ui.components.ShowProgressIndicator
+import com.backcube.economyapp.core.ui.utils.CollectEffect
 import com.backcube.economyapp.domain.utils.CurrencyIsoCode
 import com.backcube.economyapp.domain.utils.formatAsWholeThousands
-import com.backcube.economyapp.features.common.baseComponents.CustomTopBar
-import com.backcube.economyapp.features.common.ui.CustomFloatingButton
-import com.backcube.economyapp.features.common.ui.CustomListItem
-import com.backcube.economyapp.features.common.ui.ShowAlertDialog
-import com.backcube.economyapp.features.common.ui.ShowProgressIndicator
-import com.backcube.economyapp.features.common.utils.CollectEffect
 import com.backcube.economyapp.features.income.store.models.IncomeEffect
 import com.backcube.economyapp.features.income.store.models.IncomeIntent
 import com.backcube.economyapp.features.income.store.models.IncomeState
