+++ app/src/main/java/com/backcube/economyapp/ui/theme/Theme.kt
@@ -4,12 +4,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 
-//private val DarkColorScheme = darkColorScheme(
-//    primary = Purple80,
-//    secondary = PurpleGrey80,
-//    tertiary = Pink80
-//)
-
 private val LightColorScheme = lightColorScheme(
     primary = PointGreen,
     surface = SurfaceMain,
